Updated TOC to get rid of conflicts with two TOCs
The original version would allow numerous TOCs built from the same root with different orders. This could cause conflicts and doesn't make any sense with the design of the module. I added a global state for the TOC, such that each root category can have only one TOC that is saved in the database. If an editor creates two TOCs from the same root category, this is allowed, but it produces the already saved order. Everything is kept in sync.